Type Definition sptir::color::spaces::SRGBColor

type SRGBColor = RGBColor<SRGB>;

Implementations

impl SRGBColor

pub fn decode_transfer_function_from(srgb: &[u8]) -> SRGBColor

Decodes an 8-bit encoded sRGB color.

\gamma ^{-1}(u)={
    \begin{cases}{
        \frac {u}{12.92}} &= {\frac {25u}{323}}&u\leq 0.04045 \\
        \left({\tfrac {u+0.055}{1.055}}\right)^{2.4} &= \left({\tfrac {200u+11}{211}}\right)^{\frac {12}{5}}&{\text{otherwise}}
    \end{cases}
}

where $u$ is $R_{srgb}, G_{srgb}, \text{or}\ B_{srgb}$

NOTE: Input slice must have a length of at least 3.

pub fn encode_transfer_function_to(&self, srgb: &mut [u8])

Encodes a linear sRGB color to an output slice of 8-bit integers.

\gamma (u)={
    \begin{cases}
        12.92u &= {\frac {323u}{25}}&u\leq 0.0031308 \\
        1.055u^{1/2.4}-0.055 &= {\frac {211u^{\frac {5}{12}}-11}{200}}&{\text{otherwise}}
    \end{cases}
}

where $u$ is $R_{Linear}, G_{Linear}, \text{or}\ B_{Linear}$

NOTE: Output slice must have a length of at least 3.