Struct sptir::geometry::Vector3

pub struct Vector3 {
    pub x: f32,
    pub y: f32,
    pub z: f32,
}

Fields
x: f32y: f32z: f32
Implementations

impl Vector3

pub const ZERO: Vector3

pub const UP: Vector3

pub fn new(x: f32, y: f32, z: f32) -> Vector3

impl Vector3

pub fn dot(&self, other: &Self) -> f32

The dot product (or scalar product) of two N-dimensional vectors is defined as:

\mathbf{a} \cdot \mathbf{b} = \sum _{i=1}^{N}{a}_{i}{b}_{i} = {a}_{1}{b}_{1} + {a}_{2}{b}_{2} + \cdots + {a}_{n}{b}_{n}

where $\Sigma$ denotes summation.

In Euclidean space, the dot product of two vectors is equivalent to the scaled angle between them:

\mathbf{a} \cdot \mathbf{b} = \|\mathbf{a}\|\ \|\mathbf{b}\|\cos \theta

where $\|\mathbf{a}\|$ is the Euclidean norm of the vector.

As such, the dot product is often used to compute the cosine theta between two normalized vectors.

pub fn cross(&self, other: &Self) -> Vector3

The cross product (or vector product) is a binary operation on two vectors specifically in three-dimensional space, and is denoted as the symbol $\times$

Testing

\mathbf{a} \times \mathbf{b} =\left\|\mathbf{a} \right\|\left\|\mathbf{b} \right\|\sin(\theta )\ \mathbf{n}

where $\theta$ is the angle between $\mathbf{a}$ and $\mathbf{b}$ and $\mathbf{n}$ is a unit vector perpendicular to the plane containing $\mathbf{a}$ and $\mathbf{b}$.

If the vectors $\mathbf{a}$ and $\mathbf{b}$ are parallel, by the above formula, the cross product is a zero-vector.

By convention, the direction of the vector $\mathbf{n}$ is given by the right-hand rule for a right-handed coordinate system.

For a right-handed coordinate system, the cross-product boils down to:

{\begin{pmatrix}
    s_{1}\\s_{2}\\s_{3}
\end{pmatrix}} =
{\begin{pmatrix}
    a_{2}b_{3}-a_{3}b_{2} \\
    a_{3}b_{1}-a_{1}b_{3} \\
    a_{1}b_{2}-a_{2}b_{1}
\end{pmatrix}}

pub fn norm_squared(&self) -> f32

pub fn norm(&self) -> f32

Compute the Euclidean norm ($\|\mathbf{v}\|$) of the vector $\mathbf{v}$

\|\mathbf{v}\| = \sqrt{\mathbf{v} \cdot \mathbf{v}}

where $\mathbf{v} \cdot \mathbf{v}$ is the dot product.

By computing the dot product of itself, the result is $\|\mathbf{v}\| \|\mathbf{v}\|\cos\theta$, and because cosine theta between two identical vectors is 1.0, the result is the norm squared.

pub fn normalize(&self) -> Vector3

Normalizes the vector such that the 3D Euclidean norm is 1.0

\mathbf{v}_{normalized} = \frac{\mathbf{v}}{\|\mathbf{v}\|}