Struct sptir::spectrum::SpectralRange

pub struct SpectralRange {
    pub min: f32,
    pub max: f32,
    pub y_integral: f32,
}

$\int_{\lambda_{min}}^{\lambda_{max}} \overline{y}(\lambda) d\lambda$, where wavelengths can be sampled.

Fields
min: f32

$\lambda_{min}$, the minimum wavelength which can be sampled

max: f32

$\lambda_{max}$, the maximum wavelength which can be sampled

y_integral: f32

Normalizing factor

\int_{\lambda_{min}}^{\lambda_{max}} \overline{y}(\lambda) d\lambda

Implementations

impl SpectralRange

pub const VISIBLE: SpectralRange

Precomputed visible wavelength range between 360 and 840 nanometers.

pub fn new(min: f32, max: f32) -> SpectralRange

pub fn hero_to_xyz(
&self,
wl: SpectralWavelengthSamples,
flux: SpectralFlux
) -> XYZSpectrum

    \large
    \begin{aligned}
        X &= \frac{1}{C} \sum_{j=0}^C \left( \frac{1}{N} \int_{\lambda}\frac{L_{\mathrm{e}, \Omega, \lambda}(\lambda_j)}{p(\lambda_j^s)}\,{\overline {x}}(\lambda )\,d\lambda \right) \\
        Y &= \frac{1}{C} \sum_{j=0}^C \left( \frac{1}{N} \int_{\lambda}\frac{L_{\mathrm{e}, \Omega, \lambda}(\lambda_j)}{p(\lambda_j^s)}\,{\overline {y}}(\lambda )\,d\lambda \right) \\
        Z &= \frac{1}{C} \sum_{j=0}^C \left( \frac{1}{N} \int_{\lambda}\frac{L_{\mathrm{e}, \Omega, \lambda}(\lambda_j)}{p(\lambda_j^s)}\,{\overline {z}}(\lambda )\,d\lambda \right)

        \\ \text{where} \\
        N &= \int _{\lambda }I(\lambda )\,{\overline {y}}(\lambda )\,d\lambda
    \end{aligned}

and $p(\lambda_j^s)$ is the probability density function value of each sampled wavelength.

If the sampling is uniform, then $p(\lambda_j^s)$ is simply $\small\frac{1}{\lambda_{max} - \lambda_{min}}$

pub fn sample(&self, t: f32) -> SpectralWavelengthSamples

Samples a hero wavelength and NumLanes - 1 number of equidistantly spaced other wavelength samples within the defined range, using a rotation function $r_j: \Lambda \rarr \Lambda$:

    r_j \left( \lambda_{h} \right)=\left( \lambda_h - \lambda_{min} + \frac{j}{C} \overline{\lambda}\right)\textbf{mod}\ \overline{\lambda} + \lambda_{min}

where

\begin{aligned}
    \overline{\lambda} &= \lambda_{max}-\lambda_{min} \\
    \lambda_h          &= t\overline{\lambda} + \lambda_{min} \\
    j                  &= 1,\dots,C \\
    t                  &\in [0,1)
\end{aligned}