Module sptir::spectrum::units

Spectral Units (Wavelength $\lambda_h$, Flux $\Phi_{\mathbf{e},\lambda}$, Radiance $L_{\mathbf{a},\Omega,\lambda}$, Reflectance $R_{\lambda}$)

Structs

SpectralFlux
$\Phi_{\mathbf{e},\lambda}$, radiant energy in watts per metre

SpectralRadiance
$L_{\mathbf{a},\Omega,\lambda}$, surface radiance in watts per steradian per sqr-metre, per nanometer

SpectralReflectance
$R_{\lambda}$, unitless reflectance factor between 0 and 1

SpectralWavelengths
$\lambda_h$ Hero wavelength spectrum samples with PDF

Type Definitions

Lanes
Helper type to define an array of values correspoding to each sampled wavelength

NumLanes
Defines how many wavelengths should be used for HWSS